app/requirements.txt
sanic
redis
pyyaml

